--- a/notebooks/pytorch/train-mnist-cnn.ipynb
+++ b/notebooks/pytorch/train-mnist-cnn.ipynb
@@ -32,7 +32,7 @@
     "environment_file = prefix.joinpath(\"environments\", \"pytorch-example.yml\")\n",
     "\n",
     "# azure ml settings\n",
-    "environment_name = \"pytorch-iris-example\"\n",
+    "environment_name = \"pytorch-mnist-example\"\n",
     "experiment_name = \"pytorch-mnist-example\"\n",
     "compute_target = \"gpu-cluster\""
    ]