--- a/notebooks/rapids/train-airlines-hyperdrive.ipynb
+++ b/notebooks/rapids/train-airlines-hyperdrive.ipynb
@@ -35,6 +35,9 @@
     "script_dir = prefix.joinpath(\"code\", \"models\", \"rapids\")\n",
     "script_name = \"train.py\"\n",
     "\n",
+    "# environment file\n",
+    "environment_file = prefix.joinpath(\"environments\", \"rapids-example.dockerfile\")\n",
+    "\n",
     "# azure ml settings\n",
     "environment_name = \"rapids-airline-example\"\n",
     "experiment_name = \"rapids-airline-hyperdrive-example\"\n",
@@ -58,39 +61,32 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from azureml.core.dataset import Dataset\n",
+    "from azureml.core import ScriptRunConfig, Experiment, Environment, Dataset\n",
+    "from azureml.core.runconfig import MpiConfiguration\n",
     "\n",
+    "# ds = Dataset.File.from_files('https://airlinedataset.blob.core.windows.net/airline-10years/*')\n",
     "ds = Dataset.File.from_files(\n",
     "    \"https://airlinedataset.blob.core.windows.net/airline-20m/*\"\n",
     ")\n",
     "\n",
-    "# larger dataset (10 years of airline data) is also available for multi-GPU option\n",
-    "# ds = Dataset.File.from_files('https://airlinedataset.blob.core.windows.net/airline-10years/*')"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from azureml.core import ScriptRunConfig, Experiment, Environment\n",
-    "\n",
-    "arguments = [\"--data-dir\", ds.as_mount(), \"--n_bins\", 32]\n",
-    "\n",
-    "dockerfile = \"\"\"\n",
-    "FROM rapidsai/rapidsai:0.15-cuda10.2-runtime-ubuntu18.04-py3.7\n",
-    "RUN source activate rapids && \\\n",
-    "pip install azureml-sdk && \\\n",
-    "pip install azureml-dataprep && \\\n",
-    "pip install azureml-widgets\n",
-    "\"\"\"\n",
+    "arguments = [\n",
+    "    \"--data_dir\",\n",
+    "    ds.as_mount(),\n",
+    "    \"--n_bins\",\n",
+    "    32,\n",
+    "    \"--compute\",\n",
+    "    \"single-GPU\",\n",
+    "    \"--cv-folds\",\n",
+    "    1,\n",
+    "]\n",
     "\n",
     "env = Environment(environment_name)\n",
     "env.docker.enabled = True\n",
     "env.docker.base_image = None\n",
-    "env.docker.base_dockerfile = dockerfile\n",
+    "env.docker.base_dockerfile = environment_file\n",
     "env.python.user_managed_dependencies = True\n",
+    "\n",
+    "# distr_config = MpiConfiguration(process_count_per_node=2, node_count=1)\n",
     "\n",
     "src = ScriptRunConfig(\n",
     "    source_directory=script_dir,\n",
@@ -98,27 +94,28 @@
     "    arguments=arguments,\n",
     "    environment=env,\n",
     "    compute_target=compute_target,\n",
-    ")"
+    "    #  distributed_job_config=distr_config,\n",
+    ")\n",
+    "\n",
+    "# run = Experiment(ws, experiment_name).submit(src)\n",
+    "# run"
    ]
   },
   {
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Tune model hyperparameters"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
+    "## Tune model hyperparameters\n",
+    "\n",
     "Let's define the hyperparameter space to sweep over. We will tune `n_estimators`, `max_depth` and `max_features` parameters. In this example we will use random sampling to try different configuration sets of hyperparameters and maximize `Accuracy`."
    ]
   },
   {
    "cell_type": "code",
    "execution_count": null,
-   "metadata": {},
+   "metadata": {
+    "tags": []
+   },
    "outputs": [],
    "source": [
     "!pip install --upgrade azureml-train"
@@ -152,7 +149,7 @@
     "    hyperparameter_sampling=param_sampling,\n",
     "    primary_metric_name=\"Accuracy\",\n",
     "    primary_metric_goal=PrimaryMetricGoal.MAXIMIZE,\n",
-    "    max_total_runs=10,\n",
+    "    max_total_runs=5,\n",
     "    max_concurrent_runs=5,\n",
     ")\n",
     "\n",
@@ -175,9 +172,7 @@
    "cell_type": "code",
    "execution_count": null,
    "metadata": {
-    "tags": [
-     "outputPrepend"
-    ]
+    "tags": []
    },
    "outputs": [],
    "source": [