--- a/notebooks/rapids/train-airlines-hyperdrive.ipynb
+++ b/notebooks/rapids/train-airlines-hyperdrive.ipynb
@@ -15,7 +15,7 @@
    "source": [
     "from azureml.core import Workspace\n",
     "\n",
-    "ws = Workspace.from_config()\n",
+    "ws = Workspace.from_config(\"~/code/default.json\")\n",
     "ws"
    ]
   },
@@ -62,9 +62,7 @@
    "outputs": [],
    "source": [
     "from azureml.core import ScriptRunConfig, Experiment, Environment, Dataset\n",
-    "from azureml.core.runconfig import MpiConfiguration\n",
     "\n",
-    "# ds = Dataset.File.from_files('https://airlinedataset.blob.core.windows.net/airline-10years/*')\n",
     "ds = Dataset.File.from_files(\n",
     "    \"https://airlinedataset.blob.core.windows.net/airline-20m/*\"\n",
     ")\n",
@@ -86,15 +84,12 @@
     "env.docker.base_dockerfile = environment_file\n",
     "env.python.user_managed_dependencies = True\n",
     "\n",
-    "# distr_config = MpiConfiguration(process_count_per_node=2, node_count=1)\n",
-    "\n",
     "src = ScriptRunConfig(\n",
     "    source_directory=script_dir,\n",
     "    script=script_name,\n",
     "    arguments=arguments,\n",
     "    environment=env,\n",
     "    compute_target=compute_target,\n",
-    "    #  distributed_job_config=distr_config,\n",
     ")\n",
     "\n",
     "# run = Experiment(ws, experiment_name).submit(src)\n",