--- a/notebooks/rapids/train-airlines-multi.ipynb
+++ b/notebooks/rapids/train-airlines-multi.ipynb
@@ -4,7 +4,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "# Train with RAPIDS"
+    "# Train with RAPIDS with multiple GPUs"
    ]
   },
   {
@@ -15,7 +15,7 @@
    "source": [
     "from azureml.core import Workspace\n",
     "\n",
-    "ws = Workspace.from_config()\n",
+    "ws = Workspace.from_config(\"~/code/default.json\")\n",
     "ws"
    ]
   },
@@ -75,7 +75,7 @@
     "    \"--n_bins\",\n",
     "    32,\n",
     "    \"--compute\",\n",
-    "    \"single-GPU\",\n",
+    "    \"multi-GPU\",\n",
     "    \"--cv-folds\",\n",
     "    1,\n",
     "]\n",
@@ -86,7 +86,7 @@
     "env.docker.base_dockerfile = environment_file\n",
     "env.python.user_managed_dependencies = True\n",
     "\n",
-    "# distr_config = MpiConfiguration(process_count_per_node=2, node_count=1)\n",
+    "distr_config = MpiConfiguration(process_count_per_node=2, node_count=1)\n",
     "\n",
     "src = ScriptRunConfig(\n",
     "    source_directory=script_dir,\n",
@@ -94,7 +94,7 @@
     "    arguments=arguments,\n",
     "    environment=env,\n",
     "    compute_target=compute_target,\n",
-    "    #  distributed_job_config=distr_config,\n",
+    "    distributed_job_config=distr_config,\n",
     ")\n",
     "\n",
     "run = Experiment(ws, experiment_name).submit(src)\n",