--- a/notebooks/rapids/train-airlines-multi.ipynb
+++ b/notebooks/rapids/train-airlines-multi.ipynb
@@ -40,7 +40,7 @@
     "\n",
     "# azure ml settings\n",
     "environment_name = \"rapids-airline-example\"\n",
-    "experiment_name = \"rapids-airline-example\"\n",
+    "experiment_name = \"rapids-airline-multi-example\"\n",
     "compute_target = \"gpu-cluster\""
    ]
   },
@@ -64,9 +64,8 @@
     "from azureml.core import ScriptRunConfig, Experiment, Environment, Dataset\n",
     "from azureml.core.runconfig import MpiConfiguration\n",
     "\n",
-    "# ds = Dataset.File.from_files('https://airlinedataset.blob.core.windows.net/airline-10years/*')\n",
     "ds = Dataset.File.from_files(\n",
-    "    \"https://airlinedataset.blob.core.windows.net/airline-20m/*\"\n",
+    "    \"https://airlinedataset.blob.core.windows.net/airline-10years/*\"\n",
     ")\n",
     "\n",
     "arguments = [\n",