--- a/notebooks/rapids/train-airlines.ipynb
+++ b/notebooks/rapids/train-airlines.ipynb
@@ -15,7 +15,7 @@
    "source": [
     "from azureml.core import Workspace\n",
     "\n",
-    "ws = Workspace.from_config(\"~/code/default.json\")\n",
+    "ws = Workspace.from_config()\n",
     "ws"
    ]
   },