--- a/setup.ipynb
+++ b/setup.ipynb
@@ -69,7 +69,7 @@
     "\n",
     "if ct_name not in ws.compute_targets:\n",
     "    compute_config = AmlCompute.provisioning_configuration(\n",
-    "        vm_size=\"STANDARD_NC12\",\n",
+    "        vm_size=\"STANDARD_NC12_V2\",\n",
     "        max_nodes=20,\n",
     "        idle_seconds_before_scaledown=600,\n",
     "    )\n",