… code/models/rapids/rapids_csp_azure.py
… code/models/rapids/train_rapids.py
… notebooks/rapids/hpo-rapids.ipynb