done:

- rename notebooks
- run formatting
- follow format of other notebooks
- separate out single training vs hyperdrive training 
- edits to the dockerfile 
- estimator to scriptrunconfig

to do:

- get working
- switch logging to mlflow 

current issue:

running into a cuda issue:

```
The experiment failed. Finalizing run...
[2020-09-30T19:49:13.076075] TimeoutHandler __init__
[2020-09-30T19:49:13.076157] TimeoutHandler __enter__
Cleaning up all outstanding Run operations, waiting 300.0 seconds
3 items cleaning up...
Cleanup took 0.4709360599517822 seconds
[2020-09-30T19:49:13.668172] TimeoutHandler __exit__
Enter __exit__ of DatasetContextManager
Unmounting /tmp/tmpwch2sl2e.
Finishing unmounting /tmp/tmpwch2sl2e.
Exit __exit__ of DatasetContextManager
Traceback (most recent call last):
  File "train.py", line 127, in <module>
    main()
  File "train.py", line 101, in main
    X_train, X_test, y_train, y_test, _ = azure_ml.split_data(X, y, random_state=i_train_fold)
  File "/mnt/batch/tasks/shared/LS_root/jobs/default/azureml/rapids-airline-example_1601494194_1354c796/mounts/workspaceblobstore/azureml/rapids-airline-example_1601494194_1354c796/rapids_csp_azure.py", line 280, in split_data
    random_state = random_state)
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/cuml/preprocessing/model_selection.py", line 375, in train_test_split
    X = X.iloc[idxs]
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/cudf/core/indexing.py", line 180, in __getitem__
    return self._getitem_tuple_arg(arg)
  File "/opt/conda/envs/rapids/lib/python3.7/contextlib.py", line 74, in inner
    return func(*args, **kwds)
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/cudf/core/indexing.py", line 422, in _getitem_tuple_arg
    df = columns_df._gather(arg[0])
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/cudf/core/frame.py", line 488, in _gather
    self, as_column(gather_map), keep_index=keep_index
  File "cudf/_lib/copying.pyx", line 145, in cudf._lib.copying.gather
RuntimeError: after reduction step 1: cudaErrorInvalidDeviceFunction: invalid device function


Error occurred: AzureMLCompute job failed.
JobFailed: Submitted script failed with a non-zero exit code; see the driver log file for details.
```