@zronaghi 2020-09-30
@lostmygithubaccount which GPUs/compute instance did you use for this test?