@lostmygithubaccount 2020-09-30
`STANDARD_NC12` which has K80s