@zronaghi 2020-09-30
Ok, so RAPIDS is supported on Pascal and later GPU, could you test with VMs that have P40, P100 or V100s? I've added a mention in the notebook as well