@lostmygithubaccount 2020-09-30
will upgrade to NC_12_V2 (P100s) and try