@lostmygithubaccount 2020-09-30
I got it running on V100s - can you do a sanity check that it's using both GPUs to train the model correctly? 

There are a few more things to do - I should be able to get them done today - then if you could review and merge this PR we can then merge your PR back into Azure

todo:

[ ] reduce training time of runs (<8 mins) 
[ ] multi-gpu setting added to argparse
[ ] switch to mlflow logging 
[ ] persist dockerfile in environments 
[ ] add description metadata
[ ] run readme.py