@lostmygithubaccount 2020-09-30
@zronaghi new issue - any ideas? - trying out multi-GPU fails with this: 

```
---->>>> cuDF version <<<<----
 0.15.0

---->>>> cuML version <<<<----
 0.15.0

> RapidsCloudML
	Compute, Data , Model, Cloud types ('multi-GPU', 'Parquet', 'RandomForest', 'Azure')

	Multi-GPU selected

	Client information <Client: 'tcp://127.0.0.1:46225' processes=2 threads=2, memory=236.67 GB>
multi-GPU

> Loading dataset from /tmp/tmpszqbfhn9/part*.parquet

	GPU read

	Reading using dask_cudf
Starting the daemon thread to refresh tokens in background for process with pid = 284


[2020-09-30T23:51:12.772482] The experiment failed. Finalizing run...
[2020-09-30T23:51:12.772659] FinalizingInRunHistory is not called
Cleaning up all outstanding Run operations, waiting 300.0 seconds
3 items cleaning up...
Cleanup took 1.154808521270752 seconds
Enter __exit__ of DatasetContextManager
Unmounting /tmp/tmpszqbfhn9.
Finishing unmounting /tmp/tmpszqbfhn9.
Exit __exit__ of DatasetContextManager
Traceback (most recent call last):
  File "train.py", line 179, in <module>
    main()
  File "train.py", line 116, in main
    y_label="ArrDel15",
  File "/mnt/batch/tasks/shared/LS_root/jobs/default/azureml/rapids-airline-example_1601509638_65820726/mounts/workspaceblobstore/azureml/rapids-airline-example_1601509638_65820726/rapids_csp_azure.py", line 219, in load_data
    target_filename, columns=col_labels
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask_cudf/io/parquet.py", line 189, in read_parquet
    **kwargs,
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask/dataframe/io/parquet/core.py", line 239, in read_parquet
    **kwargs,
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask_cudf/io/parquet.py", line 17, in read_metadata
    meta, stats, parts, index = ArrowEngine.read_metadata(*args, **kwargs)
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask/dataframe/io/parquet/arrow.py", line 656, in read_metadata
    paths, fs, split_row_groups, gather_statistics, filters, dataset_kwargs
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask/dataframe/io/parquet/arrow.py", line 166, in _gather_metadata
    dataset, base, fns = _get_dataset_object(paths, fs, filters, dataset_kwargs)
  File "/opt/conda/envs/rapids/lib/python3.7/site-packages/dask/dataframe/io/parquet/arrow.py", line 136, in _get_dataset_object
    elif fs.isdir(paths[0]):
IndexError: list index out of range

[2020-09-30T23:51:15.774070] Finished context manager injector with Exception.

Error occurred: AzureMLCompute job failed.
JobFailed: Submitted script failed with a non-zero exit code; see the driver log file for details.
```