@zronaghi 2020-10-01
Yes the multi-GPU version expects multiple parquet files, did you see the same issue with `airline-10years` dataset? And you mean logging the models after training?