@lostmygithubaccount 2020-10-01
yes, logging the models after training. I do not see the same issue with `airline-10years` dataset, but there is a new issue - CUDA memory runs out and it also throws an error that `trained_model` is referenced before assignment