@lostmygithubaccount 2020-10-01
log file: 

```
[2020-10-01T01:11:04.939109] Entering context manager injector.
[context_manager_injector.py] Command line Options: Namespace(inject=['ProjectPythonPath:context_managers.ProjectPythonPath', 'Dataset:context_managers.Datasets', 'RunHistory:context_managers.RunHistory', 'TrackUserError:context_managers.TrackUserError', 'UserExceptions:context_managers.UserExceptions'], invocation=['train.py', '--data_dir', 'DatasetConsumptionConfig:input_6a85608e', '--n_bins', '32', '--compute', 'multi-GPU', '--cv-folds', '1'])
This is an MPI job. Rank:0
Initialize DatasetContextManager.
Starting the daemon thread to refresh tokens in background for process with pid = 279
Set Dataset input_6a85608e's target path to /tmp/tmpgmcls4e9
Enter __enter__ of DatasetContextManager
SDK version: azureml-core==1.15.0 azureml-dataprep==2.3.0. Session id: eba7fc22-707a-4976-8eaa-816bf7e94a20. Run id: rapids-airline-multi-example_1601514635_4bf8dba2.
Processing 'input_6a85608e'.
Could not refresh EngineServer credentials in rslex: No Lariat Runtime Environment is active, please initialize an Environment.
Could not refresh EngineServer credentials in rslex: No Lariat Runtime Environment is active, please initialize an Environment.
Processing dataset FileDataset
{
  "source": [
    "https://airlinedataset.blob.core.windows.net/airline-10years/*"
  ],
  "definition": [
    "GetFiles"
  ],
  "registration": {
    "id": "a13fb474-eb17-43e7-9432-da7a83bef4b3",
    "name": null,
    "version": null,
    "workspace": "Workspace.create(name='default', subscription_id='6560575d-fa06-4e7d-95fb-f962e74efd7a', resource_group='azureml-examples')"
  }
}
Mounting input_6a85608e to /tmp/tmpgmcls4e9.
Could not refresh EngineServer credentials in rslex: No Lariat Runtime Environment is active, please initialize an Environment.
Could not refresh EngineServer credentials in rslex: No Lariat Runtime Environment is active, please initialize an Environment.
Mounted input_6a85608e to /tmp/tmpgmcls4e9 as folder.
Exit __enter__ of DatasetContextManager
Entering Run History Context Manager.
Current directory:  /mnt/batch/tasks/shared/LS_root/jobs/default/azureml/rapids-airline-multi-example_1601514635_4bf8dba2/mounts/workspaceblobstore/azureml/rapids-airline-multi-example_1601514635_4bf8dba2
Preparing to call script [ train.py ] with arguments: ['--data_dir', '$input_6a85608e', '--n_bins', '32', '--compute', 'multi-GPU', '--cv-folds', '1']
After variable expansion, calling script [ train.py ] with arguments: ['--data_dir', '/tmp/tmpgmcls4e9', '--n_bins', '32', '--compute', 'multi-GPU', '--cv-folds', '1']

Script type = None

---->>>> cuDF version <<<<----
 0.15.0

---->>>> cuML version <<<<----
 0.15.0

> RapidsCloudML
	Compute, Data , Model, Cloud types ('multi-GPU', 'Parquet', 'RandomForest', 'Azure')

	Multi-GPU selected
/opt/conda/envs/rapids/lib/python3.7/site-packages/distributed/node.py:155: UserWarning: Port 8787 is already in use.
Perhaps you already have a cluster running?
Hosting the HTTP server on port 45857 instead
  http_address["port"], self.http_server.port

	Client information <Client: 'tcp://127.0.0.1:43017' processes=2 threads=2, memory=236.67 GB>
multi-GPU

> Loading dataset from /tmp/tmpgmcls4e9/part*.parquet

	GPU read

	Reading using dask_cudf
cudf/utilities/bit.hpp(19): warning: cassert: [jitify] File not found
cudf/fixed_point/fixed_point.hpp(27): warning: cassert: [jitify] File not found

	Ingestion completed in 3.732228081998983

	Dataset descriptors: (Delayed('int-b18bf357-d836-4ac7-9f28-645e53b6b5af'), 15)
	Flight_Number_Reporting_Airline    float32
Year                               float32
Quarter                            float32
Month                              float32
DayOfWeek                          float32
DOT_ID_Reporting_Airline           float32
OriginCityMarketID                 float32
DestCityMarketID                   float32
DepTime                            float32
DepDelay                           float32
DepDel15                           float32
ArrDel15                             int32
ArrDelay                           float32
AirTime                            float32
Distance                           float32
dtype: object

---->>>> Training using GPUs <<<<----


 CV fold 0 of 1


> Splitting train and test data

	X_train shape and type(Delayed('int-be7406a5-7e4a-4e16-8da8-81ea4d98220b'), 13) <class 'dask_cudf.core.DataFrame'>

	Split completed in 0.004595635000441689

> Training RandomForest estimator w/ hyper-params

	Fitting multi-GPU daskRF
/opt/conda/envs/rapids/lib/python3.7/site-packages/cuml/dask/ensemble/randomforestclassifier.py:158: UserWarning: For reproducible results in Random Forest Classifier or for almost reproducible results in Random Forest Regressor, n_streams==1 is recommended. If n_streams is > 1, results may vary due to stream/thread timing differences, even when random_seed is set
  **kwargs
/opt/conda/envs/rapids/lib/python3.7/site-packages/cuml/dask/ensemble/randomforestclassifier.py:158: UserWarning: For reproducible results in Random Forest Classifier or for almost reproducible results in Random Forest Regressor, n_streams==1 is recommended. If n_streams is > 1, results may vary due to stream/thread timing differences, even when random_seed is set
  **kwargs
distributed.worker - WARNING -  Compute Failed
Function:  _func_fit
args:      (RandomForestClassifier(split_criterion=0), [(        AirTime  DOT_ID_Reporting_Airline  ...  Quarter    Year
index                                      ...                 
0         148.0                   19977.0  ...      1.0  2018.0
1          98.0                   19977.0  ...      1.0  2018.0
2         190.0                   19977.0  ...      1.0  2018.0
3         212.0                   19977.0  ...      1.0  2018.0
4         112.0                   19977.0  ...      1.0  2018.0
...         ...                       ...  ...      ...     ...
570110     75.0                   19790.0  ...      1.0  2018.0
570113     63.0                   19790.0  ...      1.0  2018.0
570114    126.0                   19790.0  ...      1.0  2018.0
570116     91.0                   19790.0  ...      1.0  2018.0
570117    279.0                   19790.0  ...      1.0  2018.0

[455919 rows x 13 columns], index
0         0
1         0
2         0
3         0
4         0
         ..
570110    1
570
kwargs:    {}
Exception: MemoryError('std::bad_alloc: CUDA error at: ../include/rmm/mr/device/cuda_memory_resource.hpp:68: cudaErrorMemoryAllocation out of memory')

distributed.worker - WARNING -  Compute Failed
Function:  _func_fit
args:      (RandomForestClassifier(split_criterion=0), [(        AirTime  DOT_ID_Reporting_Airline  ...  Quarter    Year
index                                      ...                 
0          37.0                   20304.0  ...      1.0  2019.0
3          80.0                   20304.0  ...      1.0  2019.0
4         127.0                   20304.0  ...      1.0  2019.0
5         110.0                   20304.0  ...      1.0  2019.0
6          72.0                   20304.0  ...      1.0  2019.0
...         ...                       ...  ...      ...     ...
583977     41.0                   19690.0  ...      1.0  2019.0
583978     33.0                   19690.0  ...      1.0  2019.0
583982     22.0                   19690.0  ...      1.0  2019.0
583983     34.0                   19690.0  ...      1.0  2019.0
583984     30.0                   19690.0  ...      1.0  2019.0

[467159 rows x 13 columns], index
0         0
3         0
4         0
5         0
6         0
         ..
583977    0
583
kwargs:    {}
Exception: MemoryError('std::bad_alloc: CUDA error at: ../include/rmm/mr/device/cuda_memory_resource.hpp:68: cudaErrorMemoryAllocation out of memory')



! Error during fit 2 of 2 worker jobs failed: std::bad_alloc: CUDA error at: ../include/rmm/mr/device/cuda_memory_resource.hpp:68: cudaErrorMemoryAllocation out of memory, std::bad_alloc: CUDA error at: ../include/rmm/mr/device/cuda_memory_resource.hpp:68: cudaErrorMemoryAllocation out of memory


!error during model training: local variable 'trained_model' referenced before assignment

	Finished training in 0.0000 s
Starting the daemon thread to refresh tokens in background for process with pid = 279


[2020-10-01T01:13:43.020696] The experiment failed. Finalizing run...
[2020-10-01T01:13:43.020852] FinalizingInRunHistory is not called
Cleaning up all outstanding Run operations, waiting 300.0 seconds
3 items cleaning up...
Cleanup took 5.351221323013306 seconds
Enter __exit__ of DatasetContextManager
Unmounting /tmp/tmpgmcls4e9.
Finishing unmounting /tmp/tmpgmcls4e9.
Exit __exit__ of DatasetContextManager
Traceback (most recent call last):
  File "train.py", line 179, in <module>
    main()
  File "train.py", line 149, in main
    X_train, y_train, model_params
  File "/mnt/batch/tasks/shared/LS_root/jobs/default/azureml/rapids-airline-multi-example_1601514635_4bf8dba2/mounts/workspaceblobstore/azureml/rapids-airline-multi-example_1601514635_4bf8dba2/rapids_csp_azure.py", line 344, in train_model
    return trained_model, training_time
UnboundLocalError: local variable 'trained_model' referenced before assignment

[2020-10-01T01:13:53.706456] Finished context manager injector with Exception.
```